requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.149.0
anthropic>=0.96.0
python-dotenv==1.0.1
bcrypt==4.2.0
passlib[bcrypt]
python-jose==3.3.0
requests==2.32.3
httpx
python-multipart==0.0.9
pywebpush
